Add types for Title Props
