+++ src/ui/Title.tsx
@@ -1,0 +1,38 @@
+import { Box, getBoxProps } from './Box';
+import type { BoxProps } from './Box';
+
+const getFontWeight = (props) => {
+  if (props.size === 1) return 300;
+  return 700;
+};
+
+const getFontSize = (props) => {
+  if (props.size === 1) return 1.6;
+  return 1.2;
+};
+
+type TitleProps = BoxProps;
+
+const Title = ({ size, children, className, ...props }: TitleProps) => {
+  return (
+    <Box
+      as={`h${size}`}
+      size={size}
+      className={className}
+      // css={`
+      //   font-weight: ${getFontWeight};
+      //   font-size: ${getFontSize}rem;
+      // `}
+      {...getBoxProps(props)}
+    >
+      {children}
+    </Box>
+  );
+};
+
+Title.defaultProps = {
+  size: 2,
+};
+
+export { Title };
+export type { TitleProps };
